RobotArm-master/pyproject.toml
[project]
name = "robotarm"
version = "0.1.0"
description = "Robot Arm Controller and Observer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cv2-enumerate-cameras>=1.2.2",
    "opencv-python>=4.12.0.88",
    "openai>=1.0.0",
    "pyserial>=3.5",
    "pyyaml>=6.0.3",
    "ffmpeg-python>=0.2.0",
]

[project.scripts]
robotarm = "robotarm.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["robotarm"]

[tool.hatch.build.targets.sdist]
include = [
    "robotarm/resources/**/*",
    "robotarm/**/*.py",
]
